requirements.txt
asgiref==3.8.1
contourpy==1.3.0
cycler==0.12.1
Django==5.1.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
fonttools==4.53.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
